Separate IFakeObjectOptionsBuilder from IFakeObjectContainer
