+++ src/FakeItEasy/Core/IFakeObjectContainer.cs
@@ -7,7 +7,6 @@ namespace FakeItEasy.Core
     /// A container that can create fake objects.
     /// </summary>
     public interface IFakeObjectContainer
-        : IFakeObjectOptionsBuilder
     {
         /// <summary>
         /// Creates a dummy object of the specified type using the specified arguments if it's
